models/A2C/demo/requirements.txt
torch>=1.9.0
sentence-transformers>=2.2.0
flask>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0
nltk>=3.6.0
scikit-learn>=1.0.0
numpy>=1.21.0
requests>=2.25.0 